Add a delay to GDB server to be terminated
When "restart" request is called, the disconnectRequest() and launchRequest() will be called. In the disconnectRequest() will send a '-gdb-exit' command to exit GDB server then the launch method will called to start a new GDB server. But LaunchRequest() is not wait for GDB server is terminated completely to run and start GDB server. In this case Emulator will be hold by old GDB server, So cannot start a new GDB server.
Solution: Make a delay to wait for GDB server is terminated completely